Track Okay and new bug
