+++ Assets/Scenes/Menu/Scene1.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.44657815, g: 0.49641192, b: 0.57481617, a: 1}
+  m_IndirectSpecularColor: {r: 0.44657898, g: 0.4964133, b: 0.5748178, a: 1}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -13835,15 +13835,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   playerPrefab: {fileID: 178461042498600476, guid: 8998536374eaf447783c1b9dc6fb44c4, type: 3}
   playerCamera: {fileID: 0}
-  spawnPositions:
-  - {fileID: 189935471}
-  - {fileID: 1005631179}
-  - {fileID: 1671107324}
-  - {fileID: 1669683836}
-  minX: -450
-  maxX: 450
-  minY: -450
-  maxY: -300
 --- !u!4 &1736074454
 Transform:
   m_ObjectHideFlags: 0
