--- Assets/Scenes/Scene1.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 185c8cddaa5ec42869b3f1921802f75b
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
