+++ Assets/Scenes/Track/CitySettings.lighting
@@ -8,9 +8,9 @@ LightingSettings:
   m_PrefabAsset: {fileID: 0}
   m_Name: CitySettings
   serializedVersion: 4
-  m_GIWorkflowMode: 0
+  m_GIWorkflowMode: 1
   m_EnableBakedLightmaps: 1
-  m_EnableRealtimeLightmaps: 0
+  m_EnableRealtimeLightmaps: 1
   m_RealtimeEnvironmentLighting: 1
   m_BounceScale: 1
   m_AlbedoBoost: 1
