+++ Assets/Scenes/Track/CitySettings.lighting.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5c93eddd3c6478144a5511d1cda263c3
+guid: 6d2c37355456e1d45a38481b4c627cdf
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 4890085278179872738
