--- Assets/Scenes/Track/Multiplayer/City.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f70bf7fc67c885e4d89da97194871e1f
-timeCreated: 1499292385
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
