--- Assets/Scenes/Track/Multiplayer/CitySettings.lighting.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f1e562bc48a481d4d9c893a4d20ab3ba
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 4890085278179872738
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
