+++ Assets/Scenes/Track/Multiplayer/MT_City.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f58435ad6092951408b79f9774e065d4
+timeCreated: 1499292385
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
