--- Assets/Scenes/Track/Multiplayer/Port.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f101e7db4a9eb0c439a6bb6ead95b711
-timeCreated: 1618808085
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
