--- Assets/Scenes/Track/Multiplayer/Portsettings.lighting.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 067de68f449e874449fded64f9d43644
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 4890085278179872738
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
