--- Assets/Scenes/Track/Timer/Timer_CitySettings.lighting.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6d2c37355456e1d45a38481b4c627cdf
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 4890085278179872738
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
