--- Assets/Scenes/Track/Timer/Timer_Portsettings.lighting.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7b4020f680b56ac4cb4cd2270c0159c8
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 4890085278179872738
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
