+++ Assets/Scripts/Race/PassedChekcpoints.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc6330539332fe04f8613bd1c45ab509
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
