+++ Assets/Scripts/Race/RaceManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 155beeb3ff917e54490aedb084480af2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
