+++ Assets/Scripts/Race/TrackPossible.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class TrackPossible : MonoBehaviour
+{
+    public static List<string> MultiplayerTrack = new List<string> {"MT_City", "MT_Port"};
+    public static List<string> TimerTrack = new List<string> {"Timer_City", "Timer_Port"};
+    public static List<string> SoloTrack = new List<string> {"IA_City", "IA_Port"};
+}
