+++ Assets/Scripts/Race/TrackPossible.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1bb0152b7f72ec54bb8442ed422aee4e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
